Partially addressing #89, add error message for 403's
You get "Error in if (cnt$OK) { : argument is of length zero" if the server returns an unspecified status_code.  403 is one, associated with API permissions on otherwise valid tokens.  

Fixing this more generally would probably involve a few lines that notes if the status code doesn't fit any specified type, and returning whatever that code is for diagnostics.  Otherwise it's ambiguous.