Update read_survey.R
Fix parsing of ImportIds for column names. The previous gsub wasn't capturing properly so the names were being read as JSON. But even when fixed, sometimes there are more than one element in the JSON array for each variable. This approach reliably captures all importIDs, but does rely on JSONLite.